Moved Seconds info in seperate component
