+++ front-end/src/components/Post/GovernanceSideBar/SecondProposal.tsx
@@ -2,11 +2,10 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-import React, { useContext, useState, useEffect } from 'react';
-import { DropdownProps, Grid } from 'semantic-ui-react';
+import React, { useContext, useState } from 'react';
+import { DropdownProps } from 'semantic-ui-react';
 import styled from '@xstyled/styled-components';
 import { InjectedAccountWithMeta } from '@polkadot/extension-inject/types';
-import { formatBalance } from '@polkadot/util';
 
 import { Form } from '../../../ui-components/Form';
 import Button from '../../../ui-components/Button';
@@ -15,6 +14,7 @@ import { NotificationContext } from '../../../context/NotificationContext';
 import { NotificationStatus, LoadingStatusType } from '../../../types';
 import Loader from 'src/ui-components/Loader';
 import AccountSelectionForm from './AccountSelectionForm';
+import SecondChainInfo from './SecondChainInfo';
 
 interface Props {
 	accounts: InjectedAccountWithMeta[]
@@ -29,32 +29,6 @@ const SecondProposal = ({ className, proposalId, address, accounts, onAccountCha
 	const [loadingStatus, setLoadingStatus] = useState<LoadingStatusType>({ isLoading: false, message:'' });
 	const { queueNotification } = useContext(NotificationContext);
 	const { api, apiReady } = useContext(ApiContext);
-	const [seconds, setSeconds] = useState(0);
-	const [deposit, setDeposit] = useState('');
-
-	useEffect(() => {
-		async function getSeconds() {
-			if (!api) {
-				console.error('polkadot/api not set');
-				return;
-			}
-
-			if (!apiReady) {
-				console.error('api not ready');
-				return;
-			}
-
-			const proposals = await api.derive.democracy.proposals();
-			proposals.forEach((proposal) => {
-				if (proposal.index.toNumber() === proposalId) {
-					setSeconds(proposal.seconds.length);
-					setDeposit(formatBalance(proposal.balance));
-				}
-			});
-		}
-
-		getSeconds().catch(console.error);
-	}, [api, apiReady, proposalId]);
 
 	const secondProposal = async () => {
 		if (!api) {
@@ -81,7 +55,7 @@ const SecondProposal = ({ className, proposalId, address, accounts, onAccountCha
 
 				console.log(`Completed at block hash #${status.asInBlock.toString()}`);
 			} else {
-				if (status.isBroadcast){
+				if (status.isBroadcast) {
 					setLoadingStatus({ isLoading: true, message: 'Broadcasting the vote' });
 				}
 				console.log(`Current status: ${status.type}`);
@@ -104,22 +78,7 @@ const SecondProposal = ({ className, proposalId, address, accounts, onAccountCha
 				<div className='card'>
 					<Form standalone={false}>
 						<h4>Second</h4>
-						<Grid className={'info'} columns={3} divided>
-							<Grid.Row>
-								<Grid.Column>
-									<div><b>Deposit Amount</b></div>
-									<div>{deposit}</div>
-								</Grid.Column>
-								<Grid.Column>
-									<div><b>Seconded By</b></div>
-									{seconds ? <div>{seconds} Addresses</div> : null}
-								</Grid.Column>
-								<Grid.Column>
-									<div><b>Locked KSM</b></div>
-									<div>{seconds * parseInt(deposit.split(' ')[0])}</div>
-								</Grid.Column>
-							</Grid.Row>
-						</Grid>
+						<SecondChainInfo proposalId={proposalId} />
 						<Form.Group>
 							<Form.Field className='button-container'>
 								<Button
